requirements.txt

fastapi
uvicorn
openai
python-dotenv
jinja2
